+++ crates/cargo-contract/src/cmd/build.rs
@@ -25,6 +25,7 @@ use contract_build::{
     Network,
     OptimizationPasses,
     OutputType,
+    Target,
     UnstableFlags,
     UnstableOptions,
     Verbosity,
@@ -114,6 +115,9 @@ pub struct BuildCommand {
     /// Don't perform wasm validation checks e.g. for permitted imports.
     #[clap(long)]
     skip_wasm_validation: bool,
+    /// Which bytecode to build the contract into.
+    #[clap(long, default_value = "wasm")]
+    target: Target,
 }
 
 impl BuildCommand {
@@ -156,6 +160,7 @@ impl BuildCommand {
             lint: self.lint,
             output_type,
             skip_wasm_validation: self.skip_wasm_validation,
+            target: self.target,
         };
 
         contract_build::execute(args)
@@ -174,6 +179,8 @@ pub struct CheckCommand {
     features: Features,
     #[clap(flatten)]
     unstable_options: UnstableOptions,
+    #[clap(long, default_value = "wasm")]
+    target: Target,
 }
 
 impl CheckCommand {
@@ -196,6 +203,7 @@ impl CheckCommand {
             lint: false,
             output_type: OutputType::default(),
             skip_wasm_validation: false,
+            target: self.target,
         };
 
         contract_build::execute(args)